+++ libs/model/src/aggregates/community/GetChainNodes.query.ts
@@ -1,3 +1,4 @@
+/*
 import { TypedRequestQuery, TypedResponse, success } from '../../types';
 import { ServerControllers } from '../../routing/router';
 import { GetChainNodesResult } from 'server/controllers/server_communities_methods/get_chain_nodes';
@@ -13,3 +14,20 @@ export const getChainNodesHandler = async (
   const results = await controllers.communities.getChainNodes({});
   return success(res, results);
 };
+
+*/
+
+/*
+import { ChainNodeInstance } from '@hicommonwealth/model';
+import { ServerCommunitiesController } from '../server_communities_controller';
+
+export type GetChainNodesOptions = {};
+export type GetChainNodesResult = ChainNodeInstance[];
+
+export async function __getChainNodes(
+  this: ServerCommunitiesController,
+): Promise<GetChainNodesResult> {
+  return this.models.ChainNode.findAll();
+}
+
+*/
