--- packages/commonwealth/server/controllers/server_communities_methods/get_chain_nodes.ts
@@ -1,11 +1,0 @@
-import { ChainNodeInstance } from '@hicommonwealth/model';
-import { ServerCommunitiesController } from '../server_communities_controller';
-
-export type GetChainNodesOptions = {};
-export type GetChainNodesResult = ChainNodeInstance[];
-
-export async function __getChainNodes(
-  this: ServerCommunitiesController,
-): Promise<GetChainNodesResult> {
-  return this.models.ChainNode.findAll();
-}
